improve tests, post endpoints and model <-> schema transformation
- add endpoint for getting one post
- update PostUpdate model to not include the uid
- update endpoint for updating a post to include the uid as path parameter
- fix UUID validation
- revert serialize properties
- add classes for transforming neomodel to Pydantic schemas for use in the Config class of the schemas
- change some status codes from 400 to 404
- restructure tests once again

+++ backend/app/crud/base.py
@@ -1,9 +1,10 @@
 from datetime import datetime, timezone
 from typing import Any, Dict, Generic, Optional, Type, TypeVar, Union
+from uuid import UUID
 
 from fastapi.encoders import jsonable_encoder
-from neomodel import NodeSet, StructuredNode, UniqueIdProperty, db
-from pydantic import BaseModel
+from neomodel import NodeSet, StructuredNode, db
+from pydantic import UUID4, BaseModel
 
 ModelType = TypeVar("ModelType", bound=StructuredNode)
 CreateSchemaType = TypeVar("CreateSchemaType", bound=BaseModel)
@@ -21,13 +22,16 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         self.model = model
 
     @db.read_transaction
-    def get(self, uid: UniqueIdProperty) -> Optional[ModelType]:
+    def get(self, uid: Union[str, UUID4]) -> Optional[ModelType]:
         """
         Get one node from the model class.
         :param uid: the unique identifier of the node to get
         :return: the requested node if it exists, else None
         """
-        return self.model.nodes.get_or_none(uid=uid)
+        if isinstance(uid, UUID):
+            return self.model.nodes.get_or_none(uid=uid.hex)
+        else:
+            return self.model.nodes.get_or_none(uid=uid)
 
     @db.read_transaction
     def get_all(self) -> NodeSet:
@@ -62,7 +66,7 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
             update_data = obj_in
         else:
             update_data = obj_in.dict(exclude_unset=True)
-        for field in db_obj.serialize:
+        for field in db_obj.__properties__:
             if field in update_data:
                 setattr(db_obj, field, update_data[field])
         if hasattr(db_obj, "updated_at"):
@@ -71,12 +75,17 @@ class CRUDBase(Generic[ModelType, CreateSchemaType, UpdateSchemaType]):
         return db_obj
 
     @db.write_transaction
-    def remove(self, uid: UniqueIdProperty) -> Optional[ModelType]:
+    def remove(self, uid: Union[str, UUID4]) -> Optional[ModelType]:
         """
         Remove a node.
         :param uid: the unique identifier of the node to be removed
         :return: the removed node
         """
-        obj = self.model.nodes.get(uid=uid)
-        obj.delete()
+        if isinstance(uid, UUID):
+            obj = self.model.nodes.get_or_none(uid=uid.hex)
+        else:
+            obj = self.model.nodes.get_or_none(uid=uid)
+
+        if obj is not None:
+            obj.delete()
         return obj
